pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quantum-variational-methods-for-susy-qm"
version = "0.1.0"
description = "Code for quantum variational methods applied to supersymmetric quantum mechanics."
readme = "README.md"
requires-python = ">=3.13"
authors = [{ name = "John Kerfoot", email = "john.kerfoot@liverpool.ac.uk"}, 
            { name = "David Schaich", email = "david.schaich@liverpool.ac.uk"}, 
            { name = "Emanuele Mendicelli", email = "e.mendicelli@liverpool.ac.uk"}
            ]
license = { file = "LICENSE" }

dependencies = [
  "numpy>=2.3.4",
  "scipy>=1.16.3",
  "pennylane>=0.43.1",
  "qiskit>=2.2.3",
  "matplotlib>=3.10.7",
  "gitpython>=3.1.45",
  "pandas>=2.3.3",
]

[tool.setuptools]
package-dir = { "" = "src" }
packages = ["susy_qm"]


